Fix target interfaces in dfn of `redraw` event
The `redraw` event was defined with a `for="XRLayer"` attribute. That is not
really correct, because the even can only fire at some of the
`XRCompositionLayer` interfaces. This update adjusts the `for` attribute
accordingly, making the definition consistent with the IDL definitions of the
related `EventHandler` attributes.

(Via https://github.com/w3c/webref/issues/1216)